Removed the abstract factory
